+++ src/main/java/com/cMall/shopChat/cart/application/dto/CartItemInfo.java
@@ -1,0 +1,4 @@
+package com.cMall.shopChat.cart.application.dto;
+
+public class CartItemInfo {
+}
