+++ src/main/java/com/cMall/shopChat/cart/application/dto/request/CartItemCreateRequest.java
@@ -1,0 +1,4 @@
+package com.cMall.shopChat.cart.application.dto.request;
+
+public class CartItemCreateRequest {
+}
