+++ src/main/java/com/cMall/shopChat/cart/application/dto/response/CartCreateResponse.java
@@ -1,0 +1,4 @@
+package com.cMall.shopChat.cart.application.dto.response;
+
+public class CartCreateResponse {
+}
