--- src/main/java/com/cMall/shopChat/cart/application/dtos/CartItemInfo.java
@@ -1,4 +1,0 @@
-package com.cMall.shopChat.cart.application.dtos;
-
-public class CartItemInfo {
-}
