--- src/main/java/com/cMall/shopChat/cart/application/dtos/request/CartItemCreateRequest.java
@@ -1,4 +1,0 @@
-package com.cMall.shopChat.cart.application.dtos.request;
-
-public class CartItemCreateRequest {
-}
