--- src/main/java/com/cMall/shopChat/cart/application/dtos/request/CartItemUpdateRequest.java
@@ -1,4 +1,0 @@
-package com.cMall.shopChat.cart.application.dtos.request;
-
-public class CartItemUpdateRequest {
-}
