--- src/main/java/com/cMall/shopChat/cart/application/dtos/response/CartCreateResponse.java
@@ -1,4 +1,0 @@
-package com.cMall.shopChat.cart.application.dtos.response;
-
-public class CartCreateResponse {
-}
