+++ src/main/java/com/cMall/shopChat/order/application/dto/OrderItemInfo.java
@@ -1,0 +1,4 @@
+package com.cMall.shopChat.order.application.dto;
+
+public class OrderItemInfo {
+}
