+++ src/main/java/com/cMall/shopChat/order/application/dto/request/OrderCreateRequest.java
@@ -1,0 +1,4 @@
+package com.cMall.shopChat.order.application.dto.request;
+
+public class OrderCreateRequest {
+}
