+++ src/main/java/com/cMall/shopChat/order/application/dto/response/OrderCancelResponse.java
@@ -1,0 +1,4 @@
+package com.cMall.shopChat.order.application.dto.response;
+
+public class OrderCancelResponse {
+}
