--- src/main/java/com/cMall/shopChat/order/application/dtos/OrderItemInfo.java
@@ -1,4 +1,0 @@
-package com.cMall.shopChat.order.application.dtos;
-
-public class OrderItemInfo {
-}
