--- src/main/java/com/cMall/shopChat/order/application/dtos/request/OrderRequest.java
@@ -1,4 +1,0 @@
-package com.cMall.shopChat.order.application.dtos.request;
-
-public class OrderRequest {
-}
