--- src/main/java/com/cMall/shopChat/order/application/dtos/response/OrderCancelResponse.java
@@ -1,4 +1,0 @@
-package com.cMall.shopChat.order.application.dtos.response;
-
-public class OrderCancelResponse {
-}
