+++ src/main/java/com/cMall/shopChat/order/application/exception/OrderException.java
@@ -1,0 +1,4 @@
+package com.cMall.shopChat.order.application.exception;
+
+public class OrderException {
+}
