--- src/main/java/com/cMall/shopChat/order/application/exceptions/OrderException.java
@@ -1,4 +1,0 @@
-package com.cMall.shopChat.order.application.exceptions;
-
-public class OrderException {
-}
